Handle null PawnColumnDefs
Previously, attempting to load a table with an invalid column def would cause a NullReferenceException and prevent the UI from drawing. The error was unrecoverable except by editing the save file or introducing the missing def. This scenario could occur when a mod is removed, but also when a conditionally added def is disabled (e.g., unchecking the "Skilled Work Speed" option in Niilo's QoL).